feat: Maintain Stock

+++ digitz_erp/buying/doctype/purchase_invoice/purchase_invoice.py
@@ -239,53 +239,56 @@ class PurchaseInvoice(Document):
 
 				change_in_stock_value = new_balance_value - last_stock_ledger.balance_value
 
+			default_company = frappe.db.get_single_value("Global Settings", "default_company")
+			maintain_stock = frappe.db.get_value("Company", default_company,['maintain_stock'])
+			print("Maintain Stock :",maintain_stock)
+			if(maintain_stock == 1):
+				new_stock_ledger = frappe.new_doc("Stock Ledger")
+				new_stock_ledger.item = docitem.item
+				new_stock_ledger.item_name = docitem.item_name
+				new_stock_ledger.warehouse = docitem.warehouse
+				new_stock_ledger.posting_date = posting_date_time
 
-			new_stock_ledger = frappe.new_doc("Stock Ledger")
-			new_stock_ledger.item = docitem.item
-			new_stock_ledger.item_name = docitem.item_name
-			new_stock_ledger.warehouse = docitem.warehouse
-			new_stock_ledger.posting_date = posting_date_time
+				new_stock_ledger.qty_in = docitem.qty_in_base_unit
+				new_stock_ledger.incoming_rate = docitem.rate_in_base_unit
+				new_stock_ledger.unit = docitem.base_unit
+				new_stock_ledger.valuation_rate = valuation_rate
+				new_stock_ledger.balance_qty = new_balance_qty
+				new_stock_ledger.balance_value = new_balance_value
+				new_stock_ledger.change_in_stock_value = change_in_stock_value
+				new_stock_ledger.voucher = "Purchase Invoice"
+				new_stock_ledger.voucher_no = self.name
+				new_stock_ledger.source = "Purchase Invoice Item"
+				new_stock_ledger.source_document_id = docitem.name
+				new_stock_ledger.insert()
 
-			new_stock_ledger.qty_in = docitem.qty_in_base_unit
-			new_stock_ledger.incoming_rate = docitem.rate_in_base_unit
-			new_stock_ledger.unit = docitem.base_unit
-			new_stock_ledger.valuation_rate = valuation_rate
-			new_stock_ledger.balance_qty = new_balance_qty
-			new_stock_ledger.balance_value = new_balance_value
-			new_stock_ledger.change_in_stock_value = change_in_stock_value
-			new_stock_ledger.voucher = "Purchase Invoice"
-			new_stock_ledger.voucher_no = self.name
-			new_stock_ledger.source = "Purchase Invoice Item"
-			new_stock_ledger.source_document_id = docitem.name
-			new_stock_ledger.insert()
+				sl = frappe.get_doc("Stock Ledger", new_stock_ledger.name)
 
-			sl = frappe.get_doc("Stock Ledger", new_stock_ledger.name)
+				# If no more records for the item, update balances. otherwise it updates in the flow
+				if more_records_count_for_item==0:
 
-			# If no more records for the item, update balances. otherwise it updates in the flow
-			if more_records_count_for_item==0:
+					if frappe.db.exists('Stock Balance', {'item':docitem.item,'warehouse': docitem.warehouse}):
+						frappe.db.delete('Stock Balance',{'item': docitem.item, 'warehouse': docitem.warehouse})
 
-				if frappe.db.exists('Stock Balance', {'item':docitem.item,'warehouse': docitem.warehouse}):
-					frappe.db.delete('Stock Balance',{'item': docitem.item, 'warehouse': docitem.warehouse})
+					new_stock_balance = frappe.new_doc('Stock Balance')
+					new_stock_balance.item = docitem.item
+					new_stock_balance.item_name = docitem.item_name
+					new_stock_balance.unit = docitem.unit
+					new_stock_balance.warehouse = docitem.warehouse
+					new_stock_balance.stock_qty = new_balance_qty
+					new_stock_balance.stock_value = new_balance_value
+					new_stock_balance.valuation_rate = valuation_rate
 
-				new_stock_balance = frappe.new_doc('Stock Balance')
-				new_stock_balance.item = docitem.item
-				new_stock_balance.item_name = docitem.item_name
-				new_stock_balance.unit = docitem.unit
-				new_stock_balance.warehouse = docitem.warehouse
-				new_stock_balance.stock_qty = new_balance_qty
-				new_stock_balance.stock_value = new_balance_value
-				new_stock_balance.valuation_rate = valuation_rate
-
-				new_stock_balance.insert()
+					new_stock_balance.insert()
 
 
-				update_item_stock_balance(docitem.item)
+					update_item_stock_balance(docitem.item)
 
-			else:
-				stock_recalc_voucher.append('records',{'item': docitem.item,
-    													'warehouse': docitem.warehouse,
-                                                        'base_stock_ledger': new_stock_ledger.name
-                                                            })
+				else:
+					stock_recalc_voucher.append('records',{'item': docitem.item,
+	    													'warehouse': docitem.warehouse,
+	                                                        'base_stock_ledger': new_stock_ledger.name
+	                                                            })
 		# posting_status_doc = frappe.get_doc("Document Posting Status",{'document_type':'Purchase Invoice','document_name':self.name})
 		# posting_status_doc.stock_posted_time = datetime.now()
 		# posting_status_doc.save()
@@ -517,7 +520,7 @@ class PurchaseInvoice(Document):
 		idx +=1
 
 		# Stock Received But Not Billed - Debit - Against Trade Payable A/c
-		
+
 		gl_doc = frappe.new_doc('GL Posting')
 		gl_doc.voucher_type = "Purchase Invoice"
 		gl_doc.voucher_no = self.name
@@ -531,7 +534,7 @@ class PurchaseInvoice(Document):
 		gl_doc.insert()
 		idx +=1
 
-		# Tax - Debit - Against Trade Payable A/c	
+		# Tax - Debit - Against Trade Payable A/c
 		if self.tax_total>0:
 			gl_doc = frappe.new_doc('GL Posting')
 			gl_doc.voucher_type = "Purchase Invoice"
@@ -547,7 +550,7 @@ class PurchaseInvoice(Document):
 			idx +=1
 
 		# Round Off
-		if self.round_off!=0.00:			
+		if self.round_off!=0.00:
 			gl_doc = frappe.new_doc('GL Posting')
 			gl_doc.voucher_type = "Purchase Invoice"
 			gl_doc.voucher_no = self.name
@@ -621,7 +624,7 @@ class PurchaseInvoice(Document):
 			# posting_status_doc.payment_posted_time = datetime.now()
 			# posting_status_doc.save()
 			update_posting_status(self.doctype, self.name, 'payment_posted_time', None)
-			
+
 @frappe.whitelist()
 def get_default_payment_mode():
     default_payment_mode = frappe.db.get_value('Company', filters={'name'},fieldname='default_payment_mode_for_purchase')
